+++ package.json
@@ -30,7 +30,7 @@
     "lint-staged": "lint-staged"
   },
   "dependencies": {
-    "core-js": "^3.13.0",
+    "core-js": "^3.14.0",
     "reflect-metadata": "^0.1.13",
     "typedi": "^0.10.0",
     "uuid": "^8.3.2",
@@ -46,13 +46,13 @@
     "@rollup/plugin-commonjs": "^19.0.0",
     "@rollup/plugin-node-resolve": "^13.0.0",
     "@rollup/plugin-typescript": "^8.2.1",
-    "@tailwindcss/forms": "^0.3.2",
+    "@tailwindcss/forms": "^0.3.3",
     "@types/js-yaml": "^4.0.1",
     "@types/uuid": "^8.3.0",
-    "@typescript-eslint/eslint-plugin": "^4.25.0",
-    "@typescript-eslint/parser": "^4.25.0",
-    "@vitejs/plugin-vue": "^1.2.1",
-    "@vitejs/plugin-vue-jsx": "^1.1.4",
+    "@typescript-eslint/eslint-plugin": "^4.26.0",
+    "@typescript-eslint/parser": "^4.26.0",
+    "@vitejs/plugin-vue": "^1.2.3",
+    "@vitejs/plugin-vue-jsx": "^1.1.5",
     "@vue/babel-plugin-jsx": "^1.0.6",
     "@vue/compiler-sfc": "^3.0.11",
     "@vue/eslint-config-prettier": "^6.0.0",
@@ -62,10 +62,10 @@
     "commitizen": "^4.2.4",
     "cross-env": "^7.0.3",
     "cz-lerna-changelog": "^2.0.3",
-    "element-plus": "^1.0.2-beta.45",
-    "eslint": "^7.27.0",
+    "element-plus": "^1.0.2-beta.46",
+    "eslint": "^7.28.0",
     "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-vue": "^7.9.0",
+    "eslint-plugin-vue": "^7.10.0",
     "husky": "^6.0.0",
     "js-yaml": "^4.1.0",
     "lerna": "^4.0.0",
@@ -74,20 +74,20 @@
     "monaco-editor": "^0.24.0",
     "node-sass": "^6.0.0",
     "postcss": "^8.3.0",
-    "prettier": "^2.3.0",
+    "prettier": "^2.3.1",
     "rimraf": "^3.0.2",
-    "rollup": "^2.50.2",
+    "rollup": "^2.50.6",
     "rollup-plugin-postcss": "^4.0.0",
     "rollup-plugin-scss": "^2.6.1",
     "rollup-plugin-sizes": "^1.0.4",
     "rollup-plugin-terser": "^7.0.2",
     "rollup-plugin-vue": "^6.0.0",
-    "sass": "^1.34.0",
+    "sass": "^1.34.1",
     "semantic-release": "~15.8.0",
-    "tailwindcss": "^2.1.2",
+    "tailwindcss": "^2.1.4",
     "tslib": "^2.2.0",
     "typescript": "^4.3.2",
-    "vite": "^2.3.4",
+    "vite": "^2.3.6",
     "vitepress": "^0.14.0",
     "vue-router": "^4.0.6"
   },
