+++ packages/plugin-vue/src/index.ts
@@ -4,6 +4,7 @@ import checked from './prerender/checked'
 import events from './prerender/events'
 import value from './prerender/value'
 import text from './render/text'
+import condition from './render/condition'
 
 export default ({ datasource, prerender, render }: any) => {
   datasource('fetch', fetchDatasource)
@@ -13,5 +14,6 @@ export default ({ datasource, prerender, render }: any) => {
   prerender(checked)
   prerender(events)
 
+  render(condition, -65535)
   render(text)
 }
