+++ packages/vue/src/service/index.ts
@@ -1,37 +1,14 @@
 import {
+  assignObject,
   createServiceContainer,
-  functionalToken,
-  proxyToken,
-  ProxyMatcher,
-  datasourceToken,
-  DatasourceBuilder,
-  Hook,
-  Functional,
+  createCoreSetup,
 } from '@json2render/core'
-import { Setup } from '../types'
+import { Setup, Hook } from '../types'
 import { prerenderToken, renderToken } from '../feature/hook'
 
 export const containerBuilder = createServiceContainer()
 
-export const createSetupCollection = (container: any) => {
-  const proxySetup = (value: ProxyMatcher) => {
-    container.addValue(proxyToken, value)
-  }
-
-  const functionalSetup = (name: string, value: Functional) => {
-    container.addValue(functionalToken, {
-      name,
-      invoke: value,
-    })
-  }
-
-  const datasourceSetup = (name: string, value: DatasourceBuilder) => {
-    container.addValue(datasourceToken, {
-      type: name,
-      build: value,
-    })
-  }
-
+export const createSetup = (container: any) => {
   const prerenderSetup = (value: Hook, index?: number) => {
     container.addValue(prerenderToken, { index: index || 0, invoke: value })
   }
@@ -43,15 +20,12 @@ export const createSetupCollection = (container: any) => {
     })
   }
 
-  return {
-    proxy: proxySetup,
-    functional: functionalSetup,
-    datasource: datasourceSetup,
+  return assignObject(createCoreSetup(container), {
     prerender: prerenderSetup,
     render: renderSetup,
-  }
+  })
 }
 
 export const globalSetup: Setup = (handler) => {
-  handler(createSetupCollection(containerBuilder))
+  handler(createSetup(containerBuilder))
 }
